move logo to images folder
